Generate  server command reference docs Signed-off-by: William Tam <email.wtam@gmail.com>
